+++ src/lib.rs
@@ -31,5 +31,7 @@ pub mod http;
 pub mod manager;
 pub mod metrics;
 pub mod payload;
+pub mod test;
 pub mod targets;
 pub mod units;
+pub mod utils;